Add flags and support for setting max line length
With -l | --max-line-length you can cap the line length.  This is useful
because when matching on lines that are super long, the output gets
obscured and is completely unreadable anyway.  For this reason, the
default for this setting will be ON and at 2,000 characters.

There is also a flag for disabling the line length limit:

  -x | --no-max-line-length

This can be used to ensure that every matching line is printed